Implemented Crawler Class
The Crawler crawls the weather forecast for the next day from weather.com, The Crawler class inherits from the AsyncTask and takes in a callback instance (calls the OnCrawlCompleted method defined in it after the crawling is completed), a connectivity manager instance (for checking network) and lastly the execute method of the parent class AyncTask takes in two doubles : Latitude and Longitude of the User Location. Some utility classes and a interface are implemented.

Co-Authored-By: ScientistSidd <45243400+ScientistSidd@users.noreply.github.com>